+++ CHANGES.rst
@@ -7,7 +7,8 @@ Changelog
 
 - Fix serializer/deserializer to use block transformers.
   [cekk]
-
+- Remove html link in default configs.
+  [cekk]
 
 1.0.0 (2024-02-13)
 ------------------